Automation of Server Creation and Joining
Added a new script to the Scripts folder called TestButtonInvoke.cs.
This script is a tool that automates the process of server creation if
no servers are able to be joined and if a server is found it will join
that server. This tool works by invoking function calls from various UI
components of the NetworkLobbyManager asset in Unity.